requirements-verified.txt
# Verified, pinned versions for Python 3.9–3.11
# Use: pip install -r requirements-verified.txt

numpy==1.26.4
scipy==1.11.4
matplotlib==3.8.4
h5py==3.10.0
emcee==3.1.4
PyYAML==6.0.1
pandas==2.1.4
seaborn==0.13.2
psutil==5.9.8
rich==13.7.1
scikit-optimize==0.9.0
scikit-learn==1.3.2
plotly==5.22.0
pydantic==2.7.1
